Extract campaign discovery read model
